web/enterprice_portal/deploy/backend/requirements.txt
flask
gunicorn
pyjwt
redis
requests